+++ src/main/groovy/com/github/tonybaines/gestalt/sources/features/ExceptionOnNullValueDecorator.groovy
@@ -2,8 +2,11 @@ package com.github.tonybaines.gestalt.sources.features
 
 import com.github.tonybaines.gestalt.ConfigSource
 import com.github.tonybaines.gestalt.ConfigurationException
+import com.github.tonybaines.gestalt.Configurations
 
 import java.lang.reflect.Method
+
+import static com.github.tonybaines.gestalt.Configurations.Utils.optional
 
 class ExceptionOnNullValueDecorator implements ConfigSource {
   private final ConfigSource delegate
@@ -15,11 +18,12 @@ class ExceptionOnNullValueDecorator implements ConfigSource {
   @Override
   def lookup(List<String> path, Method method) {
     def value = delegate.lookup(path, method)
-    if (value == null) {
+    if (value == null && !optional(method)) {
       def e = new ConfigurationException(method.name, "not found in any source")
       e.setStackTrace([] as StackTraceElement[])
       throw e
     }
     else value
   }
+
 }